remove delay from PinHistory, return to its own file
